backend/requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
bcrypt==4.2.0
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.2
passlib==1.7.4
psycopg2==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.33
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
